beisi_rag/requirements.txt
faiss-cpu==1.12.0
importlib-metadata==8.0.0
jaraco.collections==5.1.0
langchain==1.0.3
langchain-community==0.4.1
langchain-openai==1.0.1
pip-chill==1.0.3
platformdirs==4.2.2
tomli==2.0.1
unicorn==2.1.4
uvicorn==0.38.0
fastapi
dotenv
dashscope